backend/backend/requirements.txt
asgiref==3.3.4
awsebcli==3.19.4
botocore==1.20.54
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
future==0.16.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
pathspec==0.5.9
pycparser==2.20
PyJWT==2.0.1
pypiwin32==223
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pywin32==300
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
